requirements.txt
matplotlib
geopy
numpy
opencv-python
Pillow
pyowm
instagrapi
requests
scikit-image